Move ungroup_grouped_df to group_indices.cpp
