--- src/group_by.cpp
@@ -1,18 +1,0 @@
-#include "pch.h"
-#include <dplyr/main.h>
-
-#include <tools/match.h>
-#include <tools/utils.h>
-
-#include <dplyr/Groups.h>
-
-using namespace Rcpp;
-using namespace dplyr;
-
-// [[Rcpp::export]]
-DataFrame ungroup_grouped_df(DataFrame df) {
-  DataFrame copy(shallow_copy(df));
-  GroupedDataFrame::strip_groups(copy);
-  set_class(copy, NaturalDataFrame::classes());
-  return copy;
-}
